Implement `PC` gate
Note that the HDL file's documentation states that if one control bit is
set then the other two control bits should be cleared, which would
suggest that the user of the gate is expected to be responsible for
making sure that only one control bit at most is set at any given time.
However, the test script contains cases where two or all three control
bits are set.

This in turn means that, in order to pass the test script, the
implementation needs to have a specific ordering of the multiplexing of
the various 16-bit bus values that could be produced by the gate. In
particular, it appears that the `reset` control bit should be checked
last, and the `load` bit needs to be checked second; this is because it
was found that having the `reset` bit checked second and the `load` bit
checked last failed the test script.

Possibly there is a way to avoid the ordering of which control bits are
set in the test script affecting the implementation of the gate, but it
does appear that if more than one control bit is allowed to be set, then
there is going to be problems arising from the ordering of the `Mux16`
gates used internally, at the very least with the chosen approach of
using a sequence of `Mux16` gates to choose which of the four possible
16-bit bus values should be stored in the internal `Register`.